requirement.txt
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==21.4.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==3.1.7
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
django-deep-serializer==0.1.3
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fonttools==4.29.1
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.6
numpy==1.22.0
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.6.1
ptyprocess==0.7.0
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.18.0
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
terminado==0.12.1
testpath==0.5.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
xkit==0.0.0
zipp==1.0.0
